+++ pages/dao/[symbol]/proposal/components/instructions/Loading.tsx
