--- pages/dao/[symbol]/proposal/components/instructions/Mango/MakeChangeMaxAccounts.tsx
@@ -1,152 +1,0 @@
-/* eslint-disable @typescript-eslint/no-non-null-assertion */
-import React, { useContext, useEffect, useState } from 'react'
-import useRealm from '@hooks/useRealm'
-import { PublicKey } from '@solana/web3.js'
-import * as yup from 'yup'
-import { isFormValid } from '@utils/formValidation'
-import {
-  UiInstruction,
-  MangoMakeChangeMaxAccountsForm,
-} from '@utils/uiTypes/proposalCreationTypes'
-import { NewProposalContext } from '../../../new'
-import useGovernanceAssets from '@hooks/useGovernanceAssets'
-import { Governance, GovernanceAccountType } from '@models/accounts'
-import { ParsedAccount } from '@models/core/accounts'
-import useWalletStore from 'stores/useWalletStore'
-import { serializeInstructionToBase64 } from '@models/serialisation'
-import Input from '@components/inputs/Input'
-import GovernedAccountSelect from '../../GovernedAccountSelect'
-import { GovernedMultiTypeAccount } from '@utils/tokens'
-import { makeChangeMaxMangoAccountsInstruction } from '@blockworks-foundation/mango-client'
-import { BN } from '@project-serum/anchor'
-
-const MakeChangeMaxAccounts = ({
-  index,
-  governance,
-}: {
-  index: number
-  governance: ParsedAccount<Governance> | null
-}) => {
-  const wallet = useWalletStore((s) => s.current)
-  const { realmInfo } = useRealm()
-  const { getGovernancesByAccountType } = useGovernanceAssets()
-  const governedProgramAccounts = getGovernancesByAccountType(
-    GovernanceAccountType.ProgramGovernance
-  ).map((x) => {
-    return {
-      governance: x,
-    }
-  })
-  const shouldBeGoverned = index !== 0 && governance
-  const programId: PublicKey | undefined = realmInfo?.programId
-  const [form, setForm] = useState<MangoMakeChangeMaxAccountsForm>({
-    governedAccount: undefined,
-    programId: programId?.toString(),
-    mangoGroupKey: undefined,
-    maxMangoAccounts: 1,
-  })
-  const [formErrors, setFormErrors] = useState({})
-  const { handleSetInstructions } = useContext(NewProposalContext)
-  const handleSetForm = ({ propertyName, value }) => {
-    setFormErrors({})
-    setForm({ ...form, [propertyName]: value })
-  }
-  const validateInstruction = async (): Promise<boolean> => {
-    const { isValid, validationErrors } = await isFormValid(schema, form)
-    setFormErrors(validationErrors)
-    return isValid
-  }
-  async function getInstruction(): Promise<UiInstruction> {
-    const isValid = await validateInstruction()
-    let serializedInstruction = ''
-    if (
-      isValid &&
-      programId &&
-      form.governedAccount?.governance?.info &&
-      wallet?.publicKey
-    ) {
-      //Mango instruction call and serialize
-      const setMaxMangoAccountsInstr = makeChangeMaxMangoAccountsInstruction(
-        form.governedAccount.governance.info.governedAccount,
-        new PublicKey(form.mangoGroupKey!),
-        form.governedAccount.governance.pubkey,
-        new BN(form.maxMangoAccounts)
-      )
-
-      serializedInstruction = serializeInstructionToBase64(
-        setMaxMangoAccountsInstr
-      )
-    }
-    const obj: UiInstruction = {
-      serializedInstruction: serializedInstruction,
-      isValid,
-      governance: form.governedAccount?.governance,
-    }
-    return obj
-  }
-  useEffect(() => {
-    handleSetForm({
-      propertyName: 'programId',
-      value: programId?.toString(),
-    })
-  }, [realmInfo?.programId])
-
-  useEffect(() => {
-    handleSetInstructions(
-      { governedAccount: form.governedAccount?.governance, getInstruction },
-      index
-    )
-  }, [form])
-  const schema = yup.object().shape({
-    bufferAddress: yup.number(),
-    governedAccount: yup
-      .object()
-      .nullable()
-      .required('Program governed account is required'),
-  })
-
-  return (
-    <>
-      {/* if you need more fields add theme to interface MangoMakeChangeMaxAccountsForm
-        then you can add inputs in here */}
-      <GovernedAccountSelect
-        label="Program"
-        governedAccounts={governedProgramAccounts as GovernedMultiTypeAccount[]}
-        onChange={(value) => {
-          handleSetForm({ value, propertyName: 'governedAccount' })
-        }}
-        value={form.governedAccount}
-        error={formErrors['governedAccount']}
-        shouldBeGoverned={shouldBeGoverned}
-        governance={governance}
-      ></GovernedAccountSelect>
-      <Input
-        label="Mango group key"
-        value={form.mangoGroupKey}
-        type="text"
-        onChange={(evt) =>
-          handleSetForm({
-            value: evt.target.value,
-            propertyName: 'mangoGroupKey',
-          })
-        }
-        error={formErrors['mangoGroupKey']}
-      />
-      <Input
-        label="Max accounts"
-        value={form.maxMangoAccounts}
-        type="number"
-        min={1}
-        onChange={(evt) =>
-          handleSetForm({
-            value: evt.target.value,
-            propertyName: 'maxMangoAccounts',
-          })
-        }
-        error={formErrors['maxMangoAccounts']}
-      />
-    </>
-  )
-}
-
-export default MakeChangeMaxAccounts
