+++ pages/dao/[symbol]/proposal/components/instructions/ProgramUpgrade.tsx
@@ -1,13 +1,12 @@
-import React, { useContext, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import useRealm from '@hooks/useRealm'
 import { PublicKey } from '@solana/web3.js'
-import * as yup from 'yup'
 import { isFormValid } from '@utils/formValidation'
 import {
   UiInstruction,
-  ProgramUpgradeForm,
+  ComponentInstructionData,
+  Instructions,
 } from '@utils/uiTypes/proposalCreationTypes'
-import { NewProposalContext } from '../../new'
 import useGovernanceAssets from '@hooks/useGovernanceAssets'
 import { Governance, GovernanceAccountType } from '@models/accounts'
 import { ParsedAccount } from '@models/core/accounts'
@@ -16,22 +15,46 @@ import { createUpgradeInstruction } from '@tools/sdk/bpfUpgradeableLoader/create
 import { serializeInstructionToBase64 } from '@models/serialisation'
 import Input from '@components/inputs/Input'
 import { debounce } from '@utils/debounce'
-import { validateBuffer } from '@utils/validations'
-import GovernedAccountSelect from '../GovernedAccountSelect'
+import { getProgramUpgradeSchema } from '@utils/validations'
 import { GovernedMultiTypeAccount } from '@utils/tokens'
 import { validateInstruction } from '@utils/instructionTools'
+import GovernedAccountSelect from '../GovernedAccountSelect'
+import TokenBalanceCard from '@components/TokenBalanceCard'
+import Button from '@components/Button'
+import { handlePropose } from 'actions/handleCreateProposal'
+import { NewProposalContext } from '../../new'
+
+export type ProgramUpgradeFormType = {
+  governedAccount: any | undefined
+  programId: any
+  bufferAddress: string
+  title?: string
+  description?: string
+}
 
 const ProgramUpgrade = ({
   index,
   governance,
+  setGovernance,
+  callback,
 }: {
   index: number
   governance: ParsedAccount<Governance> | null
+  setGovernance: any
+  callback?: any
 }) => {
   const connection = useWalletStore((s) => s.connection)
   const wallet = useWalletStore((s) => s.current)
-  const { realmInfo } = useRealm()
+
+  const [isLoading, setIsLoading] = useState(false)
+  const [formErrors, setFormErrors] = useState({})
+
+  const realmData = useRealm()
+
+  const { realmInfo } = realmData
+  const { fetchRealmGovernance } = useWalletStore((s) => s.actions)
   const { getGovernancesByAccountType } = useGovernanceAssets()
+
   const governedProgramAccounts = getGovernancesByAccountType(
     GovernanceAccountType.ProgramGovernance
   ).map((x) => {
@@ -39,22 +62,42 @@ const ProgramUpgrade = ({
       governance: x,
     }
   })
+
   const shouldBeGoverned = index !== 0 && governance
   const programId: PublicKey | undefined = realmInfo?.programId
-  const [form, setForm] = useState<ProgramUpgradeForm>({
+
+  const [form, setForm] = useState<ProgramUpgradeFormType>({
     governedAccount: undefined,
     programId: programId?.toString(),
     bufferAddress: '',
+    title: '',
+    description: '',
   })
-  const [formErrors, setFormErrors] = useState({})
-  const { handleSetInstructions } = useContext(NewProposalContext)
+
+  const schema = getProgramUpgradeSchema({ form, connection })
+
+  const [instructionsData, setInstructions] = useState<
+    ComponentInstructionData[]
+  >([{ type: Instructions.ProgramUpgrade }])
+
+  const handleSetInstructions = (val: any, index) => {
+    const newInstructions = [...instructionsData]
+
+    newInstructions[index] = { ...instructionsData[index], ...val }
+
+    setInstructions(newInstructions)
+  }
+
   const handleSetForm = ({ propertyName, value }) => {
     setFormErrors({})
     setForm({ ...form, [propertyName]: value })
   }
-  async function getInstruction(): Promise<UiInstruction> {
+
+  const getInstruction = async (): Promise<UiInstruction> => {
     const isValid = await validateInstruction({ schema, form, setFormErrors })
+
     let serializedInstruction = ''
+
     if (
       isValid &&
       programId &&
@@ -67,15 +110,19 @@ const ProgramUpgrade = ({
         form.governedAccount.governance.pubkey,
         wallet!.publicKey
       )
+
       serializedInstruction = serializeInstructionToBase64(upgradeIx)
     }
+
     const obj: UiInstruction = {
       serializedInstruction: serializedInstruction,
       isValid,
       governance: form.governedAccount?.governance,
     }
+
     return obj
   }
+
   useEffect(() => {
     handleSetForm({
       propertyName: 'programId',
@@ -91,67 +138,130 @@ const ProgramUpgrade = ({
       })
     }
   }, [form.bufferAddress])
+
   useEffect(() => {
     handleSetInstructions(
-      { governedAccount: form.governedAccount?.governance, getInstruction },
+      {
+        governedAccount: form.governedAccount?.governance,
+        getInstruction,
+      },
       index
     )
   }, [form])
-  const schema = yup.object().shape({
-    bufferAddress: yup
-      .string()
-      .test('bufferTest', 'Invalid buffer', async function (val: string) {
-        if (val) {
-          try {
-            await validateBuffer(
-              connection,
-              val,
-              form.governedAccount?.governance?.pubkey
-            )
-            return true
-          } catch (e) {
-            return this.createError({
-              message: `${e}`,
-            })
-          }
-        } else {
-          return this.createError({
-            message: `Buffer address is required`,
-          })
-        }
-      }),
-    governedAccount: yup
-      .object()
-      .nullable()
-      .required('Program governed account is required'),
-  })
+
+  const getSelectedGovernance = async () => {
+    return (await fetchRealmGovernance(
+      form.governedAccount?.governance.pubkey
+    )) as ParsedAccount<Governance>
+  }
+
+  const confirmPropose = async () => {
+    return await handlePropose({
+      getInstruction,
+      form,
+      connection,
+      callback,
+      governance: form.governedAccount?.governance,
+      realmData,
+      wallet,
+      getSelectedGovernance,
+      setIsLoading,
+    })
+  }
 
   return (
-    <>
-      <GovernedAccountSelect
-        label="Program"
-        governedAccounts={governedProgramAccounts as GovernedMultiTypeAccount[]}
-        onChange={(value) => {
-          handleSetForm({ value, propertyName: 'governedAccount' })
-        }}
-        value={form.governedAccount}
-        error={formErrors['governedAccount']}
-        shouldBeGoverned={shouldBeGoverned}
-        governance={governance}
-      ></GovernedAccountSelect>
-      <Input
-        label="Buffer address"
-        value={form.bufferAddress}
-        type="text"
-        onChange={(evt) =>
-          handleSetForm({
-            value: evt.target.value,
-            propertyName: 'bufferAddress',
-          })
-        }
-        error={formErrors['bufferAddress']}
-      />
-    </>
+    <NewProposalContext.Provider
+      value={{
+        instructionsData,
+        handleSetInstructions,
+        governance,
+        setGovernance,
+      }}
+    >
+      <div className="w-full flex md:flex-row flex-col justify-between items-start">
+        <div className="w-full flex md:mb-0 mb-20 flex-col gap-y-5 justify-start items-start md:max-w-xl rounded-xl">
+          <GovernedAccountSelect
+            noMaxWidth
+            useDefaultStyle={false}
+            className="p-2 w-full bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none max-w-xl"
+            label="Program"
+            governedAccounts={
+              governedProgramAccounts as GovernedMultiTypeAccount[]
+            }
+            onChange={(value) => {
+              handleSetForm({ value, propertyName: 'governedAccount' })
+            }}
+            value={form.governedAccount}
+            error={formErrors['governedAccount']}
+            shouldBeGoverned={shouldBeGoverned}
+            governance={governance}
+          />
+
+          <Input
+            noMaxWidth
+            useDefaultStyle={false}
+            className="p-4 w-fullb bg-bkg-3 border border-bkg-3 default-transition text-sm text-fgd-1 rounded-md focus:border-bkg-3 focus:outline-none max-w-xl"
+            wrapperClassName="my-6 w-full"
+            label="Buffer address"
+            value={form.bufferAddress}
+            type="text"
+            placeholder="Buffer address"
+            onChange={(evt) =>
+              handleSetForm({
+                value: evt.target.value,
+                propertyName: 'bufferAddress',
+              })
+            }
+            error={formErrors['bufferAddress']}
+          />
+
+          <Button
+            className="w-44 flex justify-center items-center mt-8"
+            onClick={confirmPropose}
+            isLoading={isLoading}
+            disabled={!form.bufferAddress}
+          >
+            Create proposal
+          </Button>
+        </div>
+
+        <div className="max-w-xs w-full">
+          <Input
+            noMaxWidth
+            useDefaultStyle
+            wrapperClassName="mb-6"
+            label="Title of your proposal"
+            placeholder="Title of your proposal (optional)"
+            value={form.title || ''}
+            type="text"
+            onChange={(event) =>
+              handleSetForm({
+                value: event.target.value,
+                propertyName: 'title',
+              })
+            }
+          />
+
+          <Input
+            noMaxWidth
+            useDefaultStyle
+            wrapperClassName="mb-20"
+            label="Description"
+            placeholder="Describe your proposal (optional)"
+            value={form.description}
+            type="text"
+            onChange={(event) =>
+              handleSetForm({
+                value: event.target.value,
+                propertyName: 'description',
+              })
+            }
+          />
+
+          <TokenBalanceCard />
+        </div>
+      </div>
+    </NewProposalContext.Provider>
   )
 }
