+++ pages/dao/[symbol]/proposal/components/instructions/StepThree.tsx
