+++ pages/dao/[symbol]/proposal/components/instructions/StepTwo.tsx
